added hamslice based pointers array for data nodes fields overhaul refactoring in progress

+++ nodes/egDataNodesContainer.cpp
@@ -2,13 +2,9 @@
 
 #include "egDataNodesContainer.h"
 
-int EgDataNodesContainerType::Connect(EgDatabaseType &myDB, std::string serverName = std::string("")) {
-    if (LoadLocalLayout())
-    {
-        std::cout << "ERROR: EgDataNodesContainerType can't open data nodes layout " << dataNodesTypeName << ".dnl" << std::endl;
-        return -1;
-    }
-    return 0;
+void EgDataNodesContainerType::init(EgDataNodeBlueprintType* a_dataNodeBlueprint) { // blueprint from upper layer
+    dataNodeBlueprint = a_dataNodeBlueprint;
+    LocalNodesFile-> GetLastID(dataNodeBlueprint-> blueprintName, lastNodeID); 
 }
 
 void EgDataNodesContainerType::clear() {
@@ -22,8 +18,15 @@ void EgDataNodesContainerType::clear() {
     updatedDataNodes.clear();
 }
 
-int EgDataNodesContainerType::LoadLocalLayout() {
-    return dataNodeLayout->LocalLoadLayout();
+int EgDataNodesContainerType::GetLastID() {
+    LocalNodesFile-> GetLastID(dataNodeBlueprint-> blueprintName, lastNodeID);          
+    return 0; // std::cout << "lastNodeID: " << std::dec << lastNodeID << std::endl;
+    // else
+    //    return -1;    
+}
+
+int EgDataNodesContainerType::LoadLocalBlueprint() {
+    return dataNodeBlueprint->LocalLoadBlueprint();
 }
 
 EgDataNodeType *EgDataNodesContainerType::GetNodePtrByID(EgDataNodeIDType nodeID) {
@@ -34,23 +37,24 @@ EgDataNodeType *EgDataNodesContainerType::GetNodePtrByID(EgDataNodeIDType nodeID
 }
 
 int EgDataNodesContainerType::AddDataNode(EgDataNodeType *newNode) {
-    newNode-> dataNodeID = dataNodeLayout-> getNextID();
+    newNode-> dataNodeID = 1 + lastNodeID++; // dataNodeBlueprint-> getNextID();
     dataNodes.insert(std::make_pair(newNode-> dataNodeID, newNode));
     addedDataNodes.insert(std::make_pair(newNode-> dataNodeID, newNode));
     return 0;
 }
 
-int EgDataNodesContainerType::MarkUpdatedDataNode(EgDataNodeType *theNode) {
-    auto iter = dataNodes.find(theNode->dataNodeID); // search all nodes
+int EgDataNodesContainerType::MarkUpdatedDataNode(EgDataNodeIDType nodeID) {
+    auto iter = dataNodes.find(nodeID); // search all nodes
     if (iter == dataNodes.end())
         return -1;
-    auto delIter = deletedDataNodes.find(theNode->dataNodeID); // search all nodes
+    auto delIter = deletedDataNodes.find(nodeID); // search all nodes
     if (delIter != deletedDataNodes.end())
         return -2;
-    auto addIter = addedDataNodes.find(theNode->dataNodeID); // search added nodes
+    auto addIter = addedDataNodes.find(nodeID); // search added nodes
     if (addIter != addedDataNodes.end())
         return -3;
-    updatedDataNodes.insert(std::make_pair(theNode-> dataNodeID, theNode));
+    updatedDataNodes.insert(std::make_pair(nodeID, iter->second));
+    // std::cout << "MarkUpdatedDataNode() done" << std::endl;
     return 0;
 }
 
@@ -81,7 +85,15 @@ void EgDataNodesContainerType::DeleteDataNode(const EgDataNodeIDType delID) {
 }
 
 int EgDataNodesContainerType::StoreToLocalFile() {
-    LocalNodesFile-> OpenFileToUpdate(dataNodesTypeName);
+    if ( !addedDataNodes.size() &&  !updatedDataNodes.size() && !deletedDataNodes.size()) {
+        std::cout << "DEBUG: StoreToLocalFile() all nodes maps empty: " << dataNodeBlueprint-> blueprintName << std::endl;
+        return 0;
+    }
+    bool isOk = LocalNodesFile->StartFileUpdate(dataNodeBlueprint-> blueprintName);
+    // std::cout << "StoreToLocalFile() start update" << std::endl;
+    if (isOk) {
+        isOk = LocalNodesFile-> UpdateNodesFile(dataNodeBlueprint-> blueprintName, addedDataNodes, deletedDataNodes, updatedDataNodes);
+/*
     for (auto delNodesIter : deletedDataNodes) { // 17 [first, second], <11 = dataFieldsNames.begin(); fieldsIter != dataFieldsNames.end(); ++fieldsIter) {
         // PrintEgDataNodeTypeFields(*(delNodesIter.second));
         LocalNodesFile-> DeleteDataNode(delNodesIter.second);
@@ -96,17 +108,21 @@ int EgDataNodesContainerType::StoreToLocalFile() {
     // std::cout << "===== After WriteDataNode 1 =====" << std::endl;
     for (auto newNodesIter : addedDataNodes) {
         // PrintEgDataNodeTypeFields(*(newNodesIter.second));
-        // std::cout << "StoreToLocalFile() container fieldsCount: " << std::dec << (int) ((newNodesIter.second)-> dataNodeLayout-> fieldsCount) << std::endl;
-        // std::cout << "StoreToLocalFile() container size: " << std::dec << (int) ((newNodesIter.second)-> dataFieldsContainer.dataFields.size()) << std::endl;
+        // std::cout << "StoreToLocalFile() container fieldsCount: " << std::dec << (int) ((newNodesIter.second)-> dataNodeBlueprint-> fieldsCount) << std::endl;
+        std::cout << "StoreToLocalFile() container size: " << std::dec << (int) ((newNodesIter.second)-> dataFieldsContainer.dataFields.size()) << std::endl;
         // std::cout << "newNodesIter.second: " << std::hex << (uint64_t) &((newNodesIter.second)-> dataFieldsContainer.dataFields) << std::endl;
         // std::cout << "* (newNodesIter.second) : " << std::hex << (uint64_t) &((*(newNodesIter.second)).dataFieldsContainer.dataFields) << std::endl;
         LocalNodesFile-> WriteDataNode(newNodesIter.second);
     }
+*/
     // std::cout << "===== After WriteDataNode 2 =====" << std::endl;
-    LocalNodesFile-> nodesFile.close();
-    updatedDataNodes.clear();
-    deletedDataNodes.clear();
-    addedDataNodes.clear();
+        LocalNodesFile-> nodesFile.close(); // close anyway
+        updatedDataNodes.clear(); // ?
+        deletedDataNodes.clear();
+        addedDataNodes.clear();
+    }
+    if (! isOk)
+        return -1;
     // std::cout << "===== StoreToLocalFile() exit =====" << std::endl;
     return 0;
 }
@@ -115,47 +131,55 @@ int EgDataNodesContainerType::LoadAllLocalFileNodes() {
     clear();
     EgFileOffsetType nextOffset{0};
     EgDataNodeType *newNode;
-    if (!LocalNodesFile->OpenFileToRead(dataNodesTypeName))
-    {
-        std::cout << "ERROR: loadAllLocalNodes() can't open file " << dataNodesTypeName << ".gdn" << std::endl;
+    if (!LocalNodesFile->OpenFileToRead(dataNodeBlueprint-> blueprintName)) {
+        // std::cout << "ERROR: loadAllLocalNodes() can't open file " << dataNodesTypeName << ".gdn" << std::endl;
         return -1;
     }
-    LocalNodesFile->getFirstNodeOffset(nextOffset); // FIXME check read
-    while (nextOffset)
-    {
+    LocalNodesFile-> getFirstNodeOffset(nextOffset); // FIXME check read
+    while (nextOffset) {
         // std::cout  << "getFirstNodeOffset() nextOffset = " << std::hex << nextOffset << std::endl;
-        newNode = new EgDataNodeType(dataNodeLayout);
-        newNode->dataFileOffset = nextOffset;
+        newNode = new EgDataNodeType(dataNodeBlueprint);
+        newNode-> dataFileOffset = nextOffset;
         LocalNodesFile->ReadDataNode(newNode, nextOffset);
         dataNodes.insert(std::make_pair(newNode->dataNodeID, newNode));
     }
-    LocalNodesFile->nodesFile.close();
+    LocalNodesFile-> nodesFile.close();
     return 0;
 }
 
-// ===================== Operators =======================
-
-EgDataNodesContainerType& operator << (EgDataNodesContainerType& nodesContainer, EgDataNodeType* newNode) { nodesContainer.AddDataNode(newNode); return nodesContainer; }
+EgDataNodesContainerType& EgDataNodesContainerType::operator << (EgDataNodeType* newNode) {
+    AddDataNode(newNode); 
+    return *this; 
+}
 
 // ======================== Debug ========================
 
-void PrintDataNodesContainer(EgDataNodesContainerType& container) {
-    std::cout << "EgDataNodesContainer nodes: " << std::endl;
-    for (auto nodesIter : container.dataNodes) // 17 [first, second], <11 = dataFieldsNames.begin(); fieldsIter != dataFieldsNames.end(); ++fieldsIter) {
+void EgDataNodesContainerType::PrintDataNodesContainer() {
+    std::cout << "EgDataNodesContainer nodes: " << dataNodeBlueprint-> blueprintName << std::endl;
+    for (auto nodesIter : dataNodes) // 17 [first, second], <11 = dataFieldsNames.begin(); fieldsIter != dataFieldsNames.end(); ++fieldsIter) {
         PrintEgDataNodeTypeFields(*(nodesIter.second));
-    if (container.addedDataNodes.size()) {
+    if (addedDataNodes.size()) {
          std::cout << "EgDataNodesContainer added nodes: " << std::endl;
-        for (auto nodesIter : container.addedDataNodes)
+        for (auto nodesIter : addedDataNodes)
             PrintEgDataNodeTypeFields(*(nodesIter.second));
     }
-    if (container.deletedDataNodes.size()) {
+    if (deletedDataNodes.size()) {
         std::cout << "EgDataNodesContainer deleted nodes: " << std::endl;
-        for (auto nodesIter : container.deletedDataNodes)
+        for (auto nodesIter : deletedDataNodes)
             PrintEgDataNodeTypeFields(*(nodesIter.second));
     }
-    if (container.updatedDataNodes.size()) {    
+    if (updatedDataNodes.size()) {    
         std::cout << "EgDataNodesContainer updated nodes: " << std::endl;
-        for (auto nodesIter : container.updatedDataNodes)
+        for (auto nodesIter : updatedDataNodes)
             PrintEgDataNodeTypeFields(*(nodesIter.second));
     }
+}
+
+void EgDataNodesContainerType::PrintNodesChain() {
+    if (!LocalNodesFile->OpenFileToRead(dataNodeBlueprint-> blueprintName)) {
+        std::cout << "ERROR: loadAllLocalNodes() can't open file " << dataNodeBlueprint-> blueprintName << ".gdn" << std::endl;
+        return;
+    }
+    LocalNodesFile-> PrintNodesChain();
+    LocalNodesFile-> nodesFile.close();
 }
